+++ src/index.tsx
@@ -32,7 +32,7 @@ export * from './layout/section/illustratedSection'
 export * from './layout/section/mediaSection'
 export * from './layout/section/slideSection'
 export * from './layout/section/textFieldsSection'
-export * from './layout/section/seaSection'
+export * from './layout/section/SeaSection'
 
 export * from './loadingPage'
 export * from './itemAction'