+++ src/layout/section/seaSection/SeaSection.tsx
@@ -1,19 +1,19 @@
 import React from 'react'
 
-import { SEASectionElements } from './SeaSection.style'
+import { SeaSectionElements } from './SeaSection.style'
 
 type IllustrationType = { small: string; large: string }
-export type SEASectionProps = Readonly<{
+export type SeaSectionProps = Readonly<{
   children: React.ReactNode
   leading: string
   illustationUrl: IllustrationType
 }>
 
-export const SEASection = ({ children, leading, illustationUrl }: SEASectionProps) => (
-  <SEASectionElements.Wrapper>
-    <SEASectionElements.Body>{children}</SEASectionElements.Body>
-    <SEASectionElements.Media small={illustationUrl.small} large={illustationUrl.large}>
-      <SEASectionElements.Leading as="h1">{leading}</SEASectionElements.Leading>
-    </SEASectionElements.Media>
-  </SEASectionElements.Wrapper>
+export const SeaSection = ({ children, leading, illustationUrl }: SeaSectionProps) => (
+  <SeaSectionElements.Wrapper>
+    <SeaSectionElements.Body>{children}</SeaSectionElements.Body>
+    <SeaSectionElements.Media small={illustationUrl.small} large={illustationUrl.large}>
+      <SeaSectionElements.Leading as="h1">{leading}</SeaSectionElements.Leading>
+    </SeaSectionElements.Media>
+  </SeaSectionElements.Wrapper>
 )